LPD-80345 Merge picklist tests into listTypeDefinitions test
