+++ modules/test/playwright/tests/object-web/main/listTypeDefinitions.spec.ts
@@ -34,6 +34,55 @@ export const test = mergeTests(
 );
 
 test.describe('manage picklists inside the picklists portlet', () => {
+	test('can cancel the creation of a picklist', async ({
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		await listTypeDefinitionPage.goto();
+
+		const picklistName = 'Picklist' + getRandomInt();
+
+		await listTypeDefinitionPage.addPicklistButton.click();
+
+		await listTypeDefinitionPage.modalNameInput.fill(picklistName);
+
+		await page.getByRole('button', {name: 'Cancel'}).click();
+
+		await expect(page.getByRole('link', {name: picklistName})).toBeHidden();
+	});
+
+	test('can cancel the creation of a picklist item', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await listTypeDefinitionPage.addPicklistItemButton.click();
+
+		const itemName = 'Item' + getRandomInt();
+
+		await listTypeDefinitionPage.modalNameInput.fill(itemName);
+
+		await page.getByRole('button', {name: 'Cancel'}).click();
+
+		const frameElement = await page.$('iframe');
+		const frame = await frameElement.contentFrame();
+		await frame.waitForLoadState('load');
+
+		await expect(frame.getByText('No Results Found')).toBeVisible();
+	});
+
 	test('can create a picklist', async ({
 		apiHelpers,
 		listTypeDefinitionPage,
@@ -51,6 +100,35 @@ test.describe('manage picklists inside the picklists portlet', () => {
 
 		await expect(
 			page.getByRole('link', {name: listTypeDefinition.name})
+		).toBeVisible();
+	});
+
+	test('can create picklist item', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await listTypeDefinitionPage.addPicklistItem(
+			listTypeDefinition.name,
+			itemName
+		);
+
+		await waitForAlert(page, 'The picklist item was created successfully.');
+
+		await expect(
+			listTypeDefinitionPage.getPicklistItemLinkLocator(itemName)
 		).toBeVisible();
 	});
 
@@ -81,6 +159,349 @@ test.describe('manage picklists inside the picklists portlet', () => {
 		await expect(
 			page.getByRole('link', {name: listTypeDefinition.name})
 		).toBeVisible();
+	});
+
+	test('can delete a picklist', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const picklistName = listTypeDefinition.name;
+
+		await listTypeDefinitionPage.goto();
+
+		await expect(
+			page.getByRole('link', {name: picklistName})
+		).toBeVisible();
+
+		await page
+			.getByRole('row', {name: picklistName})
+			.getByRole('button')
+			.click();
+
+		await page.getByRole('menuitem', {name: 'Delete'}).click();
+
+		await waitForAlert(page);
+
+		await expect(page.getByRole('link', {name: picklistName})).toBeHidden();
+	});
+
+	test('can delete a picklist item', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition: ListTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		const listTypeDefinitionName: string = listTypeDefinition.name;
+
+		const listTypeDefinitionEntryName = 'ListTypeDefinitionEntryName';
+
+		const listTypeDefinitionEntryKey = 'ListTypeDefinitionEntryKey';
+
+		await listTypeDefinitionPage.addPicklistItem(
+			listTypeDefinitionName,
+			listTypeDefinitionEntryName,
+			listTypeDefinitionEntryKey
+		);
+
+		const frameElement = await page.$('iframe');
+		const frame = await frameElement.contentFrame();
+		await frame.waitForLoadState('load');
+
+		await listTypeDefinitionPage.deletePicklistItem();
+		await frame.waitForLoadState('load');
+		await expect(frame.getByText('No Results Found')).toBeVisible();
+	});
+
+	test('can search for a picklist', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition1 =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition1.id,
+			type: 'listTypeDefinition',
+		});
+
+		const listTypeDefinition2 =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition2.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.waitForTimeout(500);
+
+		await page.getByPlaceholder('Search').fill(listTypeDefinition1.name);
+
+		await page.keyboard.press('Enter');
+
+		await expect(
+			page.getByRole('link', {name: listTypeDefinition1.name})
+		).toBeVisible();
+
+		await expect(
+			page.getByRole('link', {name: listTypeDefinition2.name})
+		).toBeHidden();
+	});
+
+	test('can search for a picklist item', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const itemName1 = 'ItemAlpha' + getRandomInt();
+		const itemName2 = 'ItemBeta' + getRandomInt();
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName1.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName1},
+		});
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName2.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName2},
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await page.waitForTimeout(500);
+
+		await listTypeDefinitionPage.frameLocator
+			.getByPlaceholder('Search')
+			.fill(itemName1);
+
+		await page.keyboard.press('Enter');
+
+		await expect(
+			listTypeDefinitionPage.getPicklistItemLinkLocator(itemName1)
+		).toBeVisible();
+
+		await expect(
+			listTypeDefinitionPage.getPicklistItemLinkLocator(itemName2)
+		).toBeHidden();
+	});
+
+	test('can update picklist item name', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName},
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await listTypeDefinitionPage
+			.getPicklistItemLinkLocator(itemName)
+			.click();
+
+		await listTypeDefinitionPage.modalSaveButton.waitFor({
+			state: 'visible',
+		});
+
+		const updatedName = 'UpdatedItem' + getRandomInt();
+
+		await listTypeDefinitionPage.modalNameInput.clear();
+
+		await listTypeDefinitionPage.modalNameInput.fill(updatedName);
+
+		await listTypeDefinitionPage.modalSaveButton.click();
+
+		await waitForAlert(page, 'The picklist item was updated successfully.');
+
+		await expect(
+			listTypeDefinitionPage.getPicklistItemLinkLocator(updatedName)
+		).toBeVisible();
+	});
+
+	test('can update picklist name', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const originalName = listTypeDefinition.name;
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: originalName}).click();
+
+		const updatedName = 'UpdatedPicklist' + getRandomInt();
+
+		await listTypeDefinitionPage.sidebarNameInput.clear();
+
+		await listTypeDefinitionPage.sidebarNameInput.fill(updatedName);
+
+		await listTypeDefinitionPage.sidebarSaveButton.click();
+
+		await waitForAlert(
+			page,
+			'Success:The picklist was updated successfully.'
+		);
+
+		await expect(page.getByRole('link', {name: updatedName})).toBeVisible();
+	});
+
+	test('cannot add special character for picklist item key field', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await listTypeDefinitionPage.addPicklistItemButton.click();
+
+		await listTypeDefinitionPage.modalNameInput.fill(
+			'PicklistItem' + getRandomInt()
+		);
+
+		await listTypeDefinitionPage.picklistItemKey.clear();
+
+		await listTypeDefinitionPage.picklistItemKey.fill('key!@#');
+
+		await listTypeDefinitionPage.modalSaveButton.click();
+
+		await expect(
+			page.getByText('Key must only contain letters and digits.')
+		).toBeVisible();
+	});
+
+	test('cannot leave picklist name field empty', async ({
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		await listTypeDefinitionPage.goto();
+
+		await listTypeDefinitionPage.addPicklistButton.click();
+
+		await listTypeDefinitionPage.modalSaveButton.click();
+
+		await expect(page.getByText('Required')).toBeVisible();
+	});
+
+	test('cannot update picklist item key', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName},
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await listTypeDefinitionPage
+			.getPicklistItemLinkLocator(itemName)
+			.click();
+
+		await listTypeDefinitionPage.modalSaveButton.waitFor({
+			state: 'visible',
+		});
+
+		await expect(listTypeDefinitionPage.picklistItemKey).toBeDisabled();
+	});
+
+	test('empty 	 message displayed when searching for a non-existent picklist', async ({
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		await listTypeDefinitionPage.goto();
+
+		const nonExistentName = 'NonExistentPicklist' + getRandomInt();
+
+		await page.waitForTimeout(500);
+
+		await page.getByPlaceholder('Search').fill(nonExistentName);
+
+		await page.keyboard.press('Enter');
+
+		await expect(page.getByText('No Results Found')).toBeVisible();
 	});
 
 	test('ensure picklist entry keys starting with upper case are correctly rendered in the entries', async ({
@@ -186,12 +607,12 @@ test.describe('manage picklists inside the picklists portlet', () => {
 		await waitForAlert(page, 'The picklist item was created successfully.');
 	});
 
-	test('can delete a picklist item', async ({
+	test('picklist item key field is autofilled when name field is filled', async ({
 		apiHelpers,
 		listTypeDefinitionPage,
 		page,
 	}) => {
-		const listTypeDefinition: ListTypeDefinition =
+		const listTypeDefinition =
 			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
 
 		apiHelpers.data.push({
@@ -201,29 +622,326 @@ test.describe('manage picklists inside the picklists portlet', () => {
 
 		await listTypeDefinitionPage.goto();
 
-		const listTypeDefinitionName: string = listTypeDefinition.name;
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
 
-		const listTypeDefinitionEntryName = 'ListTypeDefinitionEntryName';
+		await listTypeDefinitionPage.addPicklistItemButton.click();
 
-		const listTypeDefinitionEntryKey = 'ListTypeDefinitionEntryKey';
+		const itemName = 'Test Item Name';
 
-		await listTypeDefinitionPage.addPicklistItem(
-			listTypeDefinitionName,
-			listTypeDefinitionEntryName,
-			listTypeDefinitionEntryKey
+		await listTypeDefinitionPage.modalNameInput.fill(itemName);
+
+		await listTypeDefinitionPage.picklistItemKey.click();
+
+		await expect(listTypeDefinitionPage.picklistItemKey).not.toHaveValue(
+			''
+		);
+	});
+
+	test('updated picklist item name displayed on object entry FDS', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+		viewObjectEntriesPage,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName},
+		});
+
+		const objectFields = generateObjectFields({
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			objectFieldBusinessTypes: ['Picklist'],
+		});
+
+		const objectDefinitionAPIClient =
+			await apiHelpers.buildRestClient(ObjectDefinitionAPI);
+
+		const {body: objectDefinition} =
+			await objectDefinitionAPIClient.postObjectDefinition({
+				active: true,
+				label: {
+					en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+				},
+				name: 'ObjectDefinitionName' + getRandomInt(),
+				objectFields,
+				pluralLabel: {
+					en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+				},
+				portlet: true,
+				scope: 'company',
+				status: {
+					code: 0,
+				},
+			});
+
+		apiHelpers.data.push({
+			id: objectDefinition.id,
+			type: 'objectDefinition',
+		});
+
+		const picklistFieldLabel = objectFields[0].label['en_US'];
+
+		await viewObjectEntriesPage.goto(objectDefinition.className);
+
+		await viewObjectEntriesPage.clickAddObjectEntry(
+			objectDefinition.label['en_US']
 		);
 
-		const frameElement = await page.$('iframe');
-		const frame = await frameElement.contentFrame();
-		await frame.waitForLoadState('load');
+		await viewObjectEntriesPage.selectDropdownItem(
+			picklistFieldLabel,
+			itemName
+		);
 
-		await listTypeDefinitionPage.deletePicklistItem();
-		await frame.waitForLoadState('load');
-		await expect(frame.getByText('No Results Found')).toBeVisible();
+		await viewObjectEntriesPage.saveObjectEntryButton.click();
+
+		await waitForAlert(page);
+
+		await viewObjectEntriesPage.backButton.click();
+
+		await expect(page.getByText(itemName)).toBeVisible();
+
+		const updatedItemName = 'UpdatedItem' + getRandomInt();
+
+		await listTypeDefinitionPage.goto();
+
+		await page.getByRole('link', {name: listTypeDefinition.name}).click();
+
+		await listTypeDefinitionPage
+			.getPicklistItemLinkLocator(itemName)
+			.click();
+
+		await listTypeDefinitionPage.modalSaveButton.waitFor({
+			state: 'visible',
+		});
+
+		await listTypeDefinitionPage.modalNameInput.clear();
+
+		await listTypeDefinitionPage.modalNameInput.fill(updatedItemName);
+
+		await listTypeDefinitionPage.modalSaveButton.click();
+
+		await waitForAlert(page, 'The picklist item was updated successfully.');
+
+		await viewObjectEntriesPage.goto(objectDefinition.className);
+
+		await expect(page.getByText(updatedItemName)).toBeVisible();
+	});
+
+	test('Warn message displayed on picklist item screen for the delete action', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await listTypeDefinitionPage.addPicklistItem(
+			listTypeDefinition.name,
+			itemName
+		);
+
+		await expect(
+			listTypeDefinitionPage.frameLocator.getByRole('link', {
+				name: itemName,
+			})
+		).toBeVisible();
+
+		await listTypeDefinitionPage.frameLocator
+			.getByRole('row', {name: itemName})
+			.getByRole('button')
+			.click();
+
+		await listTypeDefinitionPage.frameLocator
+			.getByRole('menuitem', {name: 'Delete'})
+			.click();
 	});
 });
 
 test.describe('ensure picklist translation', () => {
+	test('can update picklist item translation', async ({
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+	}) => {
+		const listTypeDefinition =
+			await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+		apiHelpers.data.push({
+			id: listTypeDefinition.id,
+			type: 'listTypeDefinition',
+		});
+
+		const itemName = 'PicklistItem' + getRandomInt();
+
+		await apiHelpers.listTypeAdmin.postListTypeEntry({
+			key: itemName.toLowerCase(),
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			name_i18n: {en_US: itemName},
+		});
+
+		await listTypeDefinitionPage.goto();
+
+		await listTypeDefinitionPage.translatePicklistItem(
+			listTypeDefinition.name,
+			itemName,
+			'pt_BR'
+		);
+
+		await waitForAlert(page, 'The picklist item was updated successfully.');
+
+		await listTypeDefinitionPage
+			.getPicklistItemLinkLocator(itemName)
+			.click();
+
+		await listTypeDefinitionPage.modalSaveButton.waitFor({
+			state: 'visible',
+		});
+
+		await listTypeDefinitionPage.picklistItemTranslationButton.click();
+
+		await page
+			.getByRole('option', {
+				name: 'pt_BR language: Translated',
+			})
+			.click();
+
+		const updatedTranslation = itemName + ' updated translation';
+
+		await listTypeDefinitionPage.modalNameInput.fill(updatedTranslation);
+
+		await listTypeDefinitionPage.modalSaveButton.click();
+
+		await waitForAlert(page, 'The picklist item was updated successfully.');
+	});
+
+	test('translated picklist item name displayed on object entry page', async ({
+		accountSettingsPage,
+		apiHelpers,
+		listTypeDefinitionPage,
+		page,
+		viewObjectEntriesPage,
+	}) => {
+		try {
+			const listTypeDefinition =
+				await apiHelpers.listTypeAdmin.postRandomListTypeDefinition();
+
+			apiHelpers.data.push({
+				id: listTypeDefinition.id,
+				type: 'listTypeDefinition',
+			});
+
+			const itemName = 'PicklistItem' + getRandomInt();
+
+			await apiHelpers.listTypeAdmin.postListTypeEntry({
+				key: itemName.toLowerCase(),
+				listTypeDefinitionExternalReferenceCode:
+					listTypeDefinition.externalReferenceCode,
+				name_i18n: {en_US: itemName},
+			});
+
+			await listTypeDefinitionPage.goto();
+
+			await listTypeDefinitionPage.translatePicklistItem(
+				listTypeDefinition.name,
+				itemName,
+				'pt_BR'
+			);
+
+			await expect(listTypeDefinitionPage.basicInfoHeading).toBeVisible();
+
+			const objectFields = generateObjectFields({
+				listTypeDefinitionExternalReferenceCode:
+					listTypeDefinition.externalReferenceCode,
+				objectFieldBusinessTypes: ['Picklist'],
+			});
+
+			const objectDefinitionAPIClient =
+				await apiHelpers.buildRestClient(ObjectDefinitionAPI);
+
+			const {body: objectDefinition} =
+				await objectDefinitionAPIClient.postObjectDefinition({
+					active: true,
+					label: {
+						en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+					},
+					name: 'ObjectDefinitionName' + getRandomInt(),
+					objectFields,
+					pluralLabel: {
+						en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+					},
+					portlet: true,
+					scope: 'company',
+					status: {
+						code: 0,
+					},
+				});
+
+			apiHelpers.data.push({
+				id: objectDefinition.id,
+				type: 'objectDefinition',
+			});
+
+			await viewObjectEntriesPage.goto(objectDefinition.className);
+
+			await viewObjectEntriesPage.clickAddObjectEntry(
+				objectDefinition.label['en_US']
+			);
+
+			const picklistFieldLabel = objectFields[0].label['en_US'];
+
+			await viewObjectEntriesPage.selectDropdownItem(
+				picklistFieldLabel,
+				itemName
+			);
+
+			await viewObjectEntriesPage.saveObjectEntryButton.click();
+
+			await waitForAlert(page);
+
+			await accountSettingsPage.goToAccountSettings();
+
+			await accountSettingsPage.selectAccountLanguage({
+				languageId: 'pt_BR',
+			});
+
+			await viewObjectEntriesPage.goto(objectDefinition.className, 'pt');
+
+			await expect(
+				page.getByText(itemName + ' translated')
+			).toBeVisible();
+		}
+		finally {
+			await accountSettingsPage.selectAccountLanguage({
+				languageId: 'en_US',
+				navigate: true,
+			});
+
+			await waitForAlert(page);
+		}
+	});
+
 	test('verify if title of clear all button on multiselect picklist field is translated', async ({
 		accountSettingsPage,
 		apiHelpers,
