+++ modules/test/playwright/tests/object-web/main/objectEntries.spec.ts
@@ -2436,6 +2436,98 @@ test.describe('Manage object entries through View Object Entries', () => {
 		).toBeVisible();
 	});
 
+	test('can edit object entry with object field picklist mark as state', async ({
+		apiHelpers,
+		page,
+		viewObjectEntriesPage,
+	}) => {
+		const {listTypeDefinition, listTypeEntries} =
+			await postListTypeDefinitionListTypeEntries({
+				apiHelpers,
+				listTypeEntriesLength: 3,
+			});
+
+		const objectFields = generateObjectFields({
+			listTypeDefinitionExternalReferenceCode:
+				listTypeDefinition.externalReferenceCode,
+			objectFieldBusinessTypes: [
+				{
+					businessType: 'Picklist',
+					objectFieldSettings: [
+						{
+							name: 'defaultValueType',
+							value: 'inputAsValue' as any,
+						},
+						{
+							name: 'defaultValue',
+							value: listTypeEntries[0].name,
+						},
+					],
+					required: true,
+					state: true,
+				},
+			],
+		});
+
+		const objectDefinitionAPIClient =
+			await apiHelpers.buildRestClient(ObjectDefinitionAPI);
+
+		const {body: objectDefinition} =
+			await objectDefinitionAPIClient.postObjectDefinition({
+				active: true,
+				label: {
+					en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+				},
+				name: 'ObjectDefinitionName' + getRandomInt(),
+				objectFields,
+				pluralLabel: {
+					en_US: 'ObjectDefinitionLabel' + getRandomInt(),
+				},
+				portlet: true,
+				scope: 'company',
+				status: {
+					code: 0,
+				},
+			});
+
+		apiHelpers.data.push({
+			id: objectDefinition.id,
+			type: 'objectDefinition',
+		});
+
+		const picklistFieldLabel = objectFields[0].label['en_US'];
+		const firstEntryName = listTypeEntries[0].name_i18n['en-US'];
+		const secondEntryName = listTypeEntries[1].name_i18n['en-US'];
+
+		await viewObjectEntriesPage.goto(objectDefinition.className);
+
+		await viewObjectEntriesPage.clickAddObjectEntry(
+			objectDefinition.label['en_US']
+		);
+
+		await viewObjectEntriesPage.selectDropdownItem(
+			picklistFieldLabel,
+			firstEntryName
+		);
+
+		await viewObjectEntriesPage.saveObjectEntryButton.click();
+
+		await waitForAlert(page);
+
+		await viewObjectEntriesPage.backButton.click();
+
+		await viewObjectEntriesPage.frontendDatasetItems.first().click();
+
+		await viewObjectEntriesPage.selectDropdownItem(
+			picklistFieldLabel,
+			secondEntryName
+		);
+
+		await viewObjectEntriesPage.saveObjectEntryButton.click();
+
+		await waitForAlert(page);
+	});
+
 	test('can view all entries related to an object in the relationship field using autocomplete', async ({
 		apiHelpers,
 		page,
